manual-fork/requirements.txt
opentelemetry-sdk==1.6.2